requirements.txt
asn1crypto==1.0.1
astroid==2.3.1
attrdict==2.0.1
attrs==19.2.0
autopep8==1.4.4
base58==1.0.3
blocksec2go==1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cytoolz==0.10.0
ecdsa==0.13.3
eth-abi==2.0.0
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.7.0
hexbytes==0.2.0
idna==2.8
ipfshttpclient==0.4.12
isort==4.3.21
jsonschema==3.0.2
lazy-object-proxy==1.4.2
lru-dict==1.1.6
mccabe==0.6.1
multiaddr==0.0.8
netaddr==0.7.19
parsimonious==0.8.1
protobuf==3.10.0
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.9.0
pylint==2.4.2
pyrsistent==0.15.4
pyscard==1.9.9
python-dotenv==0.10.3
requests==2.22.0
rlp==1.1.0
six==1.12.0
toolz==0.10.0
typed-ast==1.4.0
urllib3==1.25.6
varint==1.0.2
web3==5.2.0
websockets==7.0
wrapt==1.11.2
